git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@4890 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ kern/kern/arch.h
@@ -11,7 +11,7 @@
 /*
  * Page allocation
  */
-extern size_t global_npages;
+extern uint64_t global_npages;
 
 extern struct page_stats {
     uint64_t pages_used;
